Change AC_REPOSITORY_DIR is required to false
- Fix global include/exclude prefix `$HOME` -> `~`
- Add warning for repository caching vs workflow step order

+++ tests/cache_pull.sh
@@ -8,17 +8,21 @@ export GRADLE_OPTS="-Dorg.gradle.daemon=false"
 
 # Workflow step input variables (component.yml>inputs)
 export AC_REPOSITORY_DIR="$HOME/app/workflow_data/tjrdzp35.isa/_appcircle_temp/Repository"
+#export AC_REPOSITORY_DIR=""
 export AC_CACHE_LABEL="master/app-deps"
 export AC_TOKEN_ID="x"
 
 rm -rf $HOME/.gradle
-rm -rf $AC_REPOSITORY_DIR
 
-mkdir -p $AC_REPOSITORY_DIR
-git clone $testApp $AC_REPOSITORY_DIR
+if [ ! -z $AC_REPOSITORY_DIR ]; then
+  rm -rf $AC_REPOSITORY_DIR
 
-if [ ! -L "/setup" ]; then
-  sudo ln -sf $HOME /setup
+  mkdir -p $AC_REPOSITORY_DIR
+  git clone $testApp $AC_REPOSITORY_DIR
+
+  if [ ! -L "/setup" ]; then
+    sudo ln -sf $HOME /setup
+  fi
 fi
 
 START_TIME=$SECONDS
@@ -29,5 +33,7 @@ echo "@@[section:end] Step completed: Cache Pull"
 ELAPSED_TIME=$(($SECONDS - $START_TIME))
 echo "took $ELAPSED_TIME s"
 
-cd $AC_REPOSITORY_DIR
-chmod +x ./gradlew && ./gradlew --build-cache app:assembleDebug
+if [ ! -z $AC_REPOSITORY_DIR ]; then
+  cd $AC_REPOSITORY_DIR
+  chmod +x ./gradlew && ./gradlew --build-cache app:assembleDebug
+fi
